Move the target libLLVM to rustc-dev
For running the compiler, we usually only need LLVM from `$sysroot/lib`,
which rustup will make available with `LD_LIBRARY_PATH`. We've also been
shipping LLVM in the `$target/lib` directory, which bloats the download
and installed size. The one time we do need the latter is for linking
`rustc-dev` libraries, so let's move it to that component directly.

Here are the dist sizes that I got before and after this change:

    rust-1.45.0-dev-x86_64-unknown-linux-gnu.tar.gz           182M  159M
    rust-1.45.0-dev-x86_64-unknown-linux-gnu.tar.xz           107M  91M
    rustc-1.45.0-dev-x86_64-unknown-linux-gnu.tar.gz          100M  78M
    rustc-1.45.0-dev-x86_64-unknown-linux-gnu.tar.xz          68M   53M
    rustc-dev-1.45.0-dev-x86_64-unknown-linux-gnu.tar.gz      146M  168M
    rustc-dev-1.45.0-dev-x86_64-unknown-linux-gnu.tar.xz      92M   108M

The installed size should reduce by exactly one `libLLVM.so` (~70-80M),
unless you also install `rustc-dev`, and then it should be identical.